requirements.txt
torch==2.8.0
torchvision==0.23.0
numpy==2.1.2
pandas==2.2.2
Pillow==10.4.0
tqdm==4.66.4
opencv-python-headless==4.10.0.84
